examples/yaml_config_app/requirements.txt
gcp-lolo>=2.0.0
PyYAML==6.0.2